add numpy.matrix dtype checks in compatibility across versions tests

+++ joblib/test/data/create_numpy_pickle.py
@@ -81,7 +81,7 @@ if __name__ == '__main__':
                  # np.matrix is a subclass of nd.array, here we want
                  # to verify this type of object is correctly unpickled
                  # among versions.
-                 np.matrix([0, 1, 2]),
+                 np.matrix([0, 1, 2], dtype=np.int64),
                  # unicode string with non-ascii chars
                  u"C'est l'\xe9t\xe9 !"]
